==> src/01.01.Describe_VRI.Rmd <==
---
title: "Describe Ranks"
author: "Nicole Barker"
date: "Last run: Jan 9, 2018"
output: 
  word_document:
    reference_docx: ../styles/ReportFormat_1.docx
---

## Script Abstract
Describe how the forest stand ranks are distributed across forest stands, and in relation to other forest attributes (BecZone, SpeciesGroup, Age, Height). Many similiar figures representing the same information are shown - depending on the specific question, one figure may be better than another for conveying the information. 

``` {r setup, echo=F, message=F, warning=F}
require(knitr)
opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(
  fig.path = "figures/"
)
```

``` {r load.project, message=F}
require(ProjectTemplate)
load.project()
options(digits=12)
```

## Background 

The Council of Forest Industries (COFI) has developed a risk management approach for incidental take based on avoidance of operation during the breeding season in forest stands that have high-bird value. The approach assigns a Risk Rank value to each forest stand in the interior of British Columbia based on attributes in the Vegetation Resource Inventory (VRI). The Risk Rank values are meant to represent densities of breeding birds, and are based on the combinations of BEC zone, leading tree species, stand age, and canopy height. This approach is based on expert opinion but has not been validated, and remains an elaborate hypothesis to be tested.

We wish to know whether these ranks are reliable for decision making.

## 1. First, an overview of the ranks

** VRI File - fixed late December 2017**

``` {r}
load("cache/vri_forsite_postclean.RData")
vri <- vri_forsite_postclean; rm(vri_forsite_postclean)
colnames(vri)
```

**NOTES**

* the VRI shapefile has `r nrow(vri)` rows (i.e., polygons). Each polygon is assumed to represent a stand, for simplicity's sake. 

### Eliminated unwanted data

Eliminate unranked stands, which is anything other than 1 through 6

``` {r}
vri$ForID <- with(vri, paste(BecZone, SpeciesGroup, Age, Height, sep="."))
vri <- vri[vri$Rank %in% 1:6,]
```

Restrict only to the Set of Attributes I care about

``` {r}
vri <- vri[vri$ForID %in% mastermatrix$ForID,]
```

**NOTES**

* Now the VRI shapefile has `r nrow(vri)` rows (i.e., polygons). Each polygon is assumed to represent a stand, for simplicity's sake. 

### Distribution of Ranks across BC

How are the ranks distributed across the interior of BC? i.e., how many stands are there of each rank?

``` {r}
ggplot(vri, aes(x=Rank, fill=as.factor(Rank))) + geom_bar() + myfacettheme3 + ylab("Number of Stands") + scale_fill_manual(values=palrank) + guides(fill=FALSE)
       
```

**NOTES**

* Most stands are intermediate ranks, 2-5. Some are 1. Very few are 6. 


### Distribution of ranks across forest attributer

How do the ranks correspond to BEC zone, species group, age, and height?

##### Reduce VRI file to the lookup table of one row per stand type

Since there are over 3 million polygons (stands) in the VRI file, we will see multiple forest stands with the same forest attributes and rank. To understand how ranks relate to forest attributes, we want a table of one row for each unique stand type. Then we can use the resultant table to tell us how Ranks are distributed among other forest attributes. 

``` {r}
stand_ranks <- vri[c("ForID", "Rank")]
stand_ranks <- stand_ranks[!duplicated(stand_ranks$ForID),]
```

Now we have a table of all forest stands that have one of the following values for Rank `r sort(unique(stand_ranks$Rank))`

##### Combine above table of ranks with complete exhaustive table of possible ForIDs (including unranked)

Not all stand types were ranked, either because they don't exist or because they were very rare (e.g., < 50 ha on the ground). But for the sake of completeness, here I'm creating a table with all theoretically possible ForIDs based on the various forest stand attributes that were used to rank stands. 

Merge mastermatrix with ranked stands and order the factor levels

``` {r}
stand_rank_lookup <- merge(mastermatrix, stand_ranks, by="ForID", all.x=T)
```

#### Add factor labels
``` {r}
stand_rank_lookup$BecZone <- factor(stand_rank_lookup$BecZone)
stand_rank_lookup$BecZone <- factor(stand_rank_lookup$BecZone, levels=levels(stand_rank_lookup$BecZone), labels=c("BG/PP", "BWBS", "ESSF", "ICH/MH", "IDF", "MS", "SBS/SBPS", "SWB"))
stand_rank_lookup$SpeciesGroup <- factor(stand_rank_lookup$SpeciesGroup)
levels(stand_rank_lookup$Age) <- c("3-30 yrs", "31-80 yrs", "81-120 yrs", "121-250 yrs", ">250 yrs")
levels(stand_rank_lookup$Height) <- c("<10.5 m", "10.5-19.4 m", "19.5-28.4 m", "> 28.5 m")
```

Now we have `r length(unique(stand_rank_lookup$ForID))` unique Forest Stand IDs. `r nrow(stand_rank_lookup[!is.na(stand_rank_lookup$Rank),])` of these stand types/IDs are ranked. 

# Plot relationships between Ranks and stand attributes 

#### Create second and third tables for ease of plotting
``` {r}
stand_rank_lookup$Rank <- as.numeric(as.character(stand_rank_lookup$Rank))
srlnona <- stand_rank_lookup[!is.na(stand_rank_lookup$Rank),]
srlfac <- stand_rank_lookup
srlfac$Rank <- factor(srlfac$Rank)
srlfacnona <- srlfac[!is.na(srlfac$Rank),]
```

## 1. BecZones

#### Get an overall idea of the distribution of ranks among BEC Zones

``` {r, Fig.Rank_BEC1, fig.height=5, fig.width=5, dpi=150}
ggplot(srlfacnona, aes(x=BecZone, y=Rank, col=Rank)) + geom_jitter(alpha=0.5,  position=position_jitter(height=0.1, width=0.2)) + myfacettheme2 + ylab("Stand Rank") + scale_color_manual(values=palrank) + guides(colour=FALSE)
```

####Understand how ranks are distributed among BEC Zones

Preprocessing: Reshape table to include 0 counts

``` {r}
tmpdat <- dcast(stand_rank_lookup, Rank ~ BecZone, fun.aggregate=length, value.var="Rank")
rank_counts_bec <- melt(tmpdat, id.vars="Rank")
colnames(rank_counts_bec) <- c("Rank", "BecZone", "NumberStandTypes")
rm(tmpdat)
```

``` {r, Fig.Rank_Bec2, fig.height=5, fig.width=5, dpi=150}
ggplot(stand_rank_lookup, aes(x=as.numeric(as.character(Rank)), fill=BecZone)) + geom_histogram(bins=6, col="black") + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Number of Stand Types") + scale_fill_manual(values=palbec, name = "BEC Zone") + myfacettheme3 + xlab("Stand Rank")
``` 

``` {r, Fig.Rank_Bec3, fig.height=10, fig.height=6, dpi=150}
ggplot(stand_rank_lookup, aes(x=as.numeric(as.character(Rank)), fill=BecZone)) + geom_histogram(aes(y=..density..), bins=6, col="black") + facet_wrap(~BecZone, ncol=2) + myfacettheme2 + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Proportion of Stand Types within BEC Zone") + scale_fill_manual(values=palbec, guide=FALSE) + xlab("Stand Rank")
```

``` {r, Fig.Rank_Bec4, fig.height=6, fig.width=9, dpi=200}
ggplot(rank_counts_bec, aes(x=as.numeric(as.character(Rank)), y=NumberStandTypes, fill=BecZone)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8, col="black") + scale_fill_manual(values=palbec, name="BEC Zone", guide=FALSE) + scale_x_continuous(labels=c(1:6), breaks=1:6) +ylab("Number of Stand Types") + myfacettheme3 + geom_text(aes(label=BecZone), position=position_dodge(width=0.8), vjust=0.4, hjust=-0.1, angle=90, size=3) + xlab("Stand Rank") + ylim(0, 90)
```


## 2. SpeciesGroups

*Get an overall idea of the distribution of ranks among Tree Species Groups*

``` {r, Fig.Rank_SPGrp1, fig.height=3, fig.width=7, dpi=150}
ggplot(srlnona, aes(x=SpeciesGroup, y=as.numeric(as.character(Rank)), col=SpeciesGroup)) + geom_jitter(alpha=0.5, position=position_jitter(height=0.1, width=0.2)) + myfacettheme1 + scale_y_continuous(labels=c(1:6), breaks=1:6) + scale_color_manual(values=palsg, guide=FALSE) + ylab("Stand Rank") + xlab("Dominant Tree Species")

ggplot(srlnona, aes(x=SpeciesGroup, y=as.numeric(as.character(Rank)))) + geom_jitter(alpha=0.5, position=position_jitter(height=0.05, width=0.2)) + myfacettheme1 + scale_y_continuous(labels=c(1:6), breaks=1:6) + ylab("Stand Rank") + xlab("Dominant Tree Species")
```

*Understand how ranks are distributed among Species Groups*
To account for any species groups that have no stands for a given rank. 

``` {r}
tmpdat <- dcast(stand_rank_lookup, Rank ~ SpeciesGroup, fun.aggregate=length, value.var="Rank")
rank_counts_spgrp <- melt(tmpdat, id.vars="Rank")
colnames(rank_counts_spgrp) <- c("Rank", "SpeciesGroup", "NumberStandTypes")
rm(tmpdat)
```

``` {r, Fig.Rank_SPGrp2, fig.height=7, fig.width=6}
ggplot(stand_rank_lookup, aes(x=Rank, fill=SpeciesGroup)) + geom_histogram(bins=6, col="black") + myfacettheme3 + scale_fill_manual(values=palsg, name="Dominant Tree Species") + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Number of Stand Types") + guides(fill=guide_legend(ncol=1)) + xlab("Stand Rank")
```

``` {r, Fig.Rank_SPGrp3, fig.height=6, fig.width=10, depi=200}
ggplot(rank_counts_spgrp, aes(x=Rank, y=NumberStandTypes, fill=SpeciesGroup)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8, col="black")   + myfacettheme2 + scale_fill_manual(values=palsg, name="Dominant Tree Species") + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylim(0,50) + ylab("Number of Stand Types") + xlab("Stand Rank") + guides(fill=guide_legend(ncol=6))
```

``` {r, Fig.Rank_SPGrp4, fig.height=10, fig.width=7, dpi=150}
ggplot(stand_rank_lookup, aes(x=Rank, fill=SpeciesGroup)) + geom_histogram(aes(y=..density..), bins=6, col="black") + facet_wrap(~SpeciesGroup, ncol=3) + myfacettheme2 + scale_fill_manual(values=palsg, guide=FALSE) + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Proportion of Stand Types") + xlab("Stand Rank") 
```

## 3. Species groups by BEC Zone
*Get an overall idea of the distribution of ranks among Tree Species Groups within the different Bec Zones*

``` {r, Fig.Rank_BECSPGrp1, fig.height=8, fig.width=5, dpi=150}
ggplot(srlnona, aes(x=SpeciesGroup, y=Rank, col=SpeciesGroup)) + geom_jitter(alpha=0.5, position=position_jitter(height=0.1, width=0.1)) + myfacettheme4 + facet_wrap(~BecZone, ncol=1) + scale_color_manual(values=palsg, guide=FALSE)+ scale_y_continuous(labels=c(1:6), breaks=1:6)  + ylab("Stand Rank") + xlab("Dominant Tree Species")
```

*Understand how ranks are distributed among Species Groups in the different BEC Zones*

``` {r}
tmpdat <- dcast(stand_rank_lookup, BecZone + SpeciesGroup ~ Rank, fun.aggregate=length, value.var="Rank")
rank_counts_becspgrp <- melt(tmpdat, id.vars=c("SpeciesGroup", "BecZone"))
colnames(rank_counts_becspgrp) <- c("SpeciesGroup", "BecZone", "Rank", "NumberStandTypes")
rank_counts_becspgrp <- rank_counts_becspgrp[!rank_counts_becspgrp$Rank %in% "NA",]
rm(tmpdat)
```

``` {r, Fig.Rank_BECSPGrp2, fig.height=10, fig.width=10, dpi=200}
ggplot(stand_rank_lookup, aes(x=Rank, fill=SpeciesGroup)) + geom_histogram(bins=6, col="black") + myfacettheme3 + scale_fill_manual(values=palsg, name="Dominant Tree Species") + facet_wrap(~BecZone, ncol=2) + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Number of Stand Types") + guides(fill=guide_legend(ncol=1)) + xlab("Stand Rank")
```

``` {r, Fig.Rank_BECspgrp3, fig.height=10, fig.width=7, dpi=200}
ggplot(rank_counts_becspgrp, aes(x=Rank, y=NumberStandTypes, fill=SpeciesGroup)) + geom_bar(stat="identity", col="black", position=position_dodge(width=0.8), width=0.8) + myfacettheme2 + scale_fill_manual(values=palsg, name="Dominant Tree Species") + facet_wrap(~BecZone, ncol=1) + ylab("Number of Stand types") + ylim(0,12) + guides(fill=guide_legend(ncol=7)) + geom_text(aes(label=SpeciesGroup), position=position_dodge(width=0.8), vjust=0.4, hjust=-0.1, angle=90, size=1) + xlab("Stand Rank")

ggplot(rank_counts_becspgrp, aes(x=BecZone, y=NumberStandTypes, fill=SpeciesGroup)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8, col="black") + ylim(0,11) + myfacettheme2 + scale_fill_manual(values=palsg, name="Dominant Tree Species") + facet_wrap(~Rank, ncol=1) + ylab("Number of Stand Types")+ guides(fill=guide_legend(ncol=7)) + xlab("BEC Zone") 
```

``` {r, Fig.Rank_BECSpgrp3, fig.height = 11, fig.width=8.5, dpi=300}
ggplot(rank_counts_becspgrp, aes(x=SpeciesGroup, y=NumberStandTypes, fill=BecZone)) + geom_bar(stat="identity", col="black", position=position_dodge(width=0.8), width=0.8) + ylim(0,12) + myfacettheme1 + scale_fill_manual(values=palbec, name="BEC Zone") + facet_wrap(~Rank, ncol=1) + ylab("Number of Stand Types") + xlab("Dominant Tree Species")

ggplot(rank_counts_becspgrp, aes(x=SpeciesGroup, y=NumberStandTypes, fill=Rank)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8) + ylim(0,12) + myfacettheme4 +  facet_wrap(~BecZone, ncol=1) + ylab("Number of Stand Types") + scale_fill_manual(values=palrank, guide=FALSE) + geom_text(aes(label=Rank), position=position_dodge(width=0.8), vjust=-0.3, hjust=0.3, angle=0, size=2) + xlab("Dominant Tree Species")

ggplot(rank_counts_becspgrp, aes(x=SpeciesGroup, y=NumberStandTypes, fill=Rank)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8) + ylim(0,12) + myfacettheme1 +  facet_wrap(~BecZone, ncol=1) + ylab("Number of Stand Types") + scale_fill_manual(values=palrank, name="Stand Rank") +  xlab("Dominant Tree Species") + guides(fill=guide_legend(ncol=6))
```

## 4. Age

#### Get an overall idea of the distribution of ranks among forest stand age classes

``` {r, Fig.Rank_age1, fig.height=3, fig.width=3, dpi=150}
ggplot(srlnona, aes(x=Age, y=Rank, col=as.factor(Rank))) + geom_jitter(alpha=0.5,  position=position_jitter(height=0.1, width=0.2)) + myfacettheme2 + scale_y_continuous(breaks=c(1:6), labels=c(1:6)) + ylab("Stand Rank") + scale_colour_manual(values = palrank, guide=FALSE) + xlab("Stand Age")
```

####Understand how ranks are distributed among forest age classes

Preprocessing: Reshape table to include 0 counts

``` {r}
tmpdat <- dcast(stand_rank_lookup, Rank ~ Age, fun.aggregate=length, value.var="Rank")
rank_counts_age <- melt(tmpdat, id.vars="Rank")
colnames(rank_counts_age) <- c("Rank", "Age", "NumberStandTypes")
rank_counts_age <- rank_counts_age[!is.na(rank_counts_age$Rank),]
```

``` {r, Fig.Rank_age2, fig.height=5, fig.width=5, dpi=150}
ggplot(stand_rank_lookup, aes(x=Rank, fill=Age)) + geom_histogram(bins=6, col="black") + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Number of Stand Types") + scale_fill_brewer(palette="YlOrRd", name = "Age Class") + myfacettheme3 + xlab("Stand Rank")
``` 

``` {r, Fig.Rank_age3, fig.height=8, width=4, dpi=150}
ggplot(stand_rank_lookup, aes(x=Rank, fill=Age)) + geom_histogram(aes(y=..density..), bins=6, col="black") + facet_wrap(~Age, ncol=1) + myfacettheme2 + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Proportion of Stand Types") + scale_fill_brewer(palette="YlOrRd", guide=FALSE) + xlab("Stand Rank")
```

``` {r, Fig.Rank_age4, fig.height=6, fig.width=9, dpi=200}
ggplot(rank_counts_age, aes(x=Rank, y=NumberStandTypes, fill=Age)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8, col="black") +  scale_fill_brewer(palette="YlOrRd", name="Age Class", guide=FALSE) + scale_x_continuous(labels=c(1:6), breaks=1:6) +ylab("Number of Stand Types") + myfacettheme3 + geom_text(aes(label=Age), position=position_dodge(width=0.8), vjust=0.4, hjust=-0.1, angle=90, size=3) + ylim(0,140) + xlab("Stand Rank")
```


## 5. Height

*Get an overall idea of the distribution of ranks among forest height classes*

``` {r, Fig.Rank_height1, fig.height=5, fig.width=7, dpi=150}
ggplot(stand_rank_lookup, aes(x=Height, y=Rank)) + geom_jitter(alpha=0.5, position=position_jitter(height=0.1, width=0.2)) + myfacettheme1 + scale_y_continuous(labels=c(1:6), breaks=1:6) +  ylab("Stand Rank") + xlab("Height Class")
```

*Understand how ranks are distributed among forest height classes*
To account for any species groups that have no stands for a given rank. 

``` {r}
tmpdat <- dcast(stand_rank_lookup, Rank ~ Height, fun.aggregate=length, value.var="Rank")
rank_counts_height <- melt(tmpdat, id.vars="Rank")
colnames(rank_counts_height) <- c("Rank", "Height", "NumberStandTypes")
rank_counts_height <- rank_counts_height[!is.na(rank_counts_height$Rank),]
```

``` {r, Fig.Rank_height2, fig.height=5, fig.width=8}
ggplot(stand_rank_lookup, aes(x=Rank, fill=Height)) + geom_histogram(bins=6, col="black") + myfacettheme2 + scale_fill_manual(values=palht, name="Height Class") + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Number of Stand Types") + xlab("Stand Rank")

ggplot(rank_counts_height, aes(x=Rank, y=NumberStandTypes, fill=Height)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8, col="black")   + myfacettheme2 + scale_fill_manual(values=palht, name="Height Class") + scale_x_continuous(labels=c(1:6), breaks=1:6)  + ylab("Number of Stand Types") + geom_text(aes(label=Height), position=position_dodge(width=0.8), vjust=0.2, hjust=-0.1, angle=90, size=3) + xlab("Stand Rank")
```

``` {r, Fig.Rank_height3, fig.height=10, fig.width=7}
ggplot(stand_rank_lookup, aes(x=Rank, fill=Height)) + geom_histogram(aes(y=..density..), bins=6, col="black") + facet_wrap(~Height, ncol=1) + myfacettheme2 + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Proportion of Stand Types") + scale_fill_manual(values=palht, guide=FALSE) + xlab("Stand Rank")
```

## 6. Height by Age
*Get an overall idea of the distribution of ranks among height classes for given age classes*

``` {r, Fig.Rank_ageht1, fig.height=6, fig.width=4, dpi=150}
ggplot(srlnona, aes(x=Height, y=Rank, col=Age)) + geom_jitter(alpha=0.2, position=position_jitter(height=0.2, width=0.2)) + myfacettheme4 + scale_y_continuous(labels=c(1:6), breaks=1:6) + guides(col=guide_legend(ncol=1)) + ylab("Stand Rank") + xlab("Height Class") + scale_color_brewer(palette="YlOrRd", name = "Age Class")

ggplot(srlfacnona, aes(x=Height, y=Rank, fill=Rank)) + geom_jitter(shape=21, alpha=0.5,col="black", position=position_jitter(height=0.1, width=0.1)) + myfacettheme4 + facet_wrap(~Age, ncol=1) +  guides(col=guide_legend(ncol=1)) + ylab("Stand Rank") + xlab("Height Class") + scale_fill_manual(values=palrank)
```

*Understand how ranks are distributed among height classes for different age classes*

``` {r}
tmpdat <- dcast(stand_rank_lookup, Height + Age ~ Rank, fun.aggregate=length, drop=FALSE, value.var="Rank")
rank_counts_ageht <- melt(tmpdat[1:8], id.vars=c("Age", "Height"))
colnames(rank_counts_ageht) <- c("Age", "Height", "Rank", "NumberStandTypes")
```

``` {r, Fig.Rank_ageht2, fig.height=10, fig.width=5}
ggplot(stand_rank_lookup, aes(x=Rank, fill=Height)) + geom_histogram(bins=6, col="black") + myfacettheme3 + scale_fill_manual(values=palht, name="Height Class") + facet_wrap(~Age, ncol=1) + scale_x_continuous(labels=c(1:6), breaks=1:6) + ylab("Number of Stand Types") + xlab("Stand Rank")

ggplot(rank_counts_ageht, aes(x=Rank, y=NumberStandTypes, fill=Height)) + geom_bar(stat="identity", col="black", position=position_dodge(width=0.8), width=0.8) + myfacettheme2 + scale_fill_manual(values=palht, name="Height Class") + facet_wrap(~Age, ncol=1) + ylab("Number of Stand types") + xlab("Stand Rank")

ggplot(rank_counts_ageht, aes(x=Age, y=NumberStandTypes, fill=Height)) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8, col="black") + myfacettheme2 + scale_fill_manual(values=palht, guide=FALSE) + facet_wrap(~Rank, ncol=1) + ylab("Number of Stand Types") + geom_text(aes(label=Height), position=position_dodge(width=0.8), vjust=0.0, hjust=-0.1, angle=90, size=3) + xlab("Age Class")

ggplot(rank_counts_ageht, aes(x=Height, y=NumberStandTypes, fill=Age)) + geom_bar(stat="identity", col="black", position=position_dodge(width=0.8), width=0.8) + myfacettheme2 + scale_fill_brewer(palette="YlOrRd", guide=FALSE) + facet_wrap(~Rank, ncol=1) + ylab("Number of Stand Types") + geom_text(aes(label=Age), position=position_dodge(width=0.8), vjust=0.4, hjust=-0.1, angle=90, size=3) + xlab("Height Class")
```


``` {r Fig.Rank_ageht3, fig.width=8, fig.height=5}
ggplot(stand_rank_lookup, aes(x=Age, y=Rank, fill=Height)) +  geom_boxplot(col="black", alpha=0.2, outlier.colour=NA, position=position_dodge(width=0.8)) + geom_point(shape=21, alpha=0.5, position=position_jitterdodge(dodge.width = 0.8, jitter.height =0.1, jitter.width = 0.3))+ scale_fill_manual(values=palht) + scale_color_manual(values=palht) + myfacettheme3 + ylab("Stand Rank") + xlab("Stand Age") + ylim("1", "2", "3", "4", "5", "6")

ggplot(stand_rank_lookup, aes(x=Age, y=Rank, fill=Height)) +   geom_point(shape=21, alpha=0.5, position=position_jitterdodge(dodge.width = 0.8, jitter.height =0.1, jitter.width = 0.3))+ scale_fill_manual(values=palht) + scale_color_manual(values=palht) + myfacettheme3 + ylab("Stand Rank") + xlab("Stand Age")
```

``` {r Fig.Rank_ageht4, fig.width=8, fig.height=6, dpi=150}
ggplot(rank_counts_ageht, aes(x=Height, y=NumberStandTypes, fill=as.factor(Rank))) + geom_bar(stat="identity", col="black", position=position_dodge(width=0.8), width=0.8) + myfacettheme3 +  facet_wrap(~Age, ncol=1) + ylab("Number of Stand Types") + scale_fill_manual(values=palrank, guide=FALSE) + geom_text(aes(label=Rank), position=position_dodge(width=0.8), vjust=-0.3, hjust=0.3, angle=0, size=3)

ggplot(rank_counts_ageht, aes(x=Height, y=NumberStandTypes, fill=as.factor(Rank))) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8) + myfacettheme3 +  facet_wrap(~Age, ncol=1) + ylab("Number of Stand Types") + scale_fill_manual(values=palrank, guide=FALSE) + geom_text(aes(label=Rank), position=position_dodge(width=0.8), vjust=-0.3, hjust=0.3, angle=0, size=3)

ggplot(rank_counts_ageht, aes(x=Height, y=NumberStandTypes, fill=as.factor(Rank))) + geom_bar(stat="identity", position=position_dodge(width=0.8), width=0.8) + myfacettheme2 +  facet_wrap(~Age, ncol=1) + ylab("Number of Stand Types") + scale_fill_manual(values=palrank, name="Stand Rank") + guides(fill=guide_legend(ncol=6))

```
